assignment details moved to components in sections
