+++ Web Ui/src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import GestionTareas from "./sections/Assignments/AssignmentsManager";
-import AssignmentDetail from "./sections/Assignments/AssignmentDetail";
+import AssignmentDetail from "./sections/Assignments/components/AssignmentDetail";
 import {GithubAPIAdapter} from "./TDDCycles-Visualization/repository/GithubAPIAdapter";//Revisar el cambio por puerto
 import TDDChartPage from "./sections/TDDCycles-Visualization/TDDChartPage";
